requirements.txt
langchain
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
python-dotenv
sentence-transformers